fix cli test name to avoid test detection
